@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,19 +27,19 @@ dependencies:
   image_picker: ^0.7.3
   video_player: ^1.0.1
   webview_flutter: ^1.0.7
-  provider: ^4.3.3
+  provider: ^5.0.0
   device_info: ^2.0.0
   date_format: ^1.0.9
   flutter_localizations:
     sdk: flutter
   shared_preferences: ^2.0.5
-  flutter_swiper: ^1.1.6
-  web_socket_channel: ^1.2.0
+  web_socket_channel: ^2.0.0
   fl_chart: ^0.20.1
-  rive: ^0.6.8
+  rive: ^0.7.1
   fluttertoast: ^8.0.3
   permission_handler: ^6.1.1
   sqflite: ^2.0.0+3
+  carousel_slider:  ^4.0.0-nullsafety.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
